🚚 Move: 그룹 api 폴더 변경
