+++ src/services/axiosInstance.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 // 임시 토큰 설정
-const TOKEN = 'gaeun@gmail.com';
+const TOKEN = 'jongwon@gmail.com';
 
 export const axiosInstance = axios.create({
   baseURL: `${import.meta.env.VITE_SERVER_URL}`,
